- Mostly complete, onto first debugging phase
Implemented priority queue and supporting data structures from the
ground up. Decided to go with a binomial heap for the base
implementation. I've put an IComparable mechanism into place to solve
the issue of having a slow DequeueMax. There will be two heaps under the
hood of the priority queue, each sorted in opposite order, but
containing the same elements where corresponding elements have
references to each other to allow for quick removal.